+++ CHANGELOG.md
@@ -25,6 +25,9 @@ Please file a bug if you notice a violation of semantic versioning.
 ### Removed
 ### Fixed
 - Add .licenserc.yaml to gem package
+- Handling of GFM fenced code blocks in CHANGELOG.md
+- Handling of nested list items in CHANGELOG.md
+- Handling of blank lines around all headings in CHANGELOG.md
 ### Security
 
 ## [1.1.19] - 2025-09-14